backend/pyproject.toml
[project]
name = "capstone-backend"
version = "0.2.0"
description = "Investment Analytics API - FastAPI Backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "supabase>=2.4.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.36",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "prophet>=1.1.5",
]

[project.optional-dependencies]
ml = [
    "tensorflow>=2.15.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "core", "data_engine", "analytics"]
